requirements.txt
Django==1.9.1
django-cors-headers==1.1.0
django-extensions==1.6.1
django-filter==0.11.0
django-taggit==0.17.6
django-taggit-serializer==0.1.5
djangorestframework==3.3.2
gunicorn==19.4.5
Pillow==3.1.0
psycopg2==2.6.1
python-memcached==1.57
redis==2.10.5
six==1.10.0
sorl-thumbnail==12.4a1
virtualenv==13.1.2
wheel==0.26.0
